Fixed the bolding issue.
I fixed the bolding issue by using html. The first try didn't work as I was trying to use Markdown inside of a HTML code snippet. 

This is a prime example of why you always should check the "Preview" tab before Commiting changes. This was an error on my behalf.